修复编译错误
- 修复 VelocityGrimReverse 中的类型转换问题
- 修复 ModuleCobwebFly 中的属性访问问题
- 修复 ModuleFollowBot 中的方法引用和碰撞检测
- 修复 ModuleItemImageReplace 中的正则表达式转义
- 修复 ModuleSkinChanger 中的lambda参数类型
- 修复 MixinDrawContextItemReplace 中的lambda参数声明
- 修复 ThemeManager 默认主题资源缺失问题

+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/combat/velocity/mode/VelocityGrimReverse.kt
@@ -65,12 +65,12 @@ internal object VelocityGrimReverse : VelocityMode("GrimReverse") {
                     val noiseX = randomNoise()
                     val noiseZ = randomNoise()
 
-                    knockback.x = -(knockback.x * (modifier + noiseX)).toFloat()
-                    knockback.z = -(knockback.z * (modifier + noiseZ)).toFloat()
+                    knockback.x = -(knockback.x * (modifier + noiseX))
+                    knockback.z = -(knockback.z * (modifier + noiseZ))
 
                     if (invertVertical) {
                         val noiseY = randomNoise(divider = 1.5)
-                        knockback.y = -(knockback.y * (modifier + noiseY)).toFloat()
+                        knockback.y = -(knockback.y * (modifier + noiseY))
                     }
 
                     NoFallBlink.waitUntilGround = true
